Split functions into utils
