+++ frontend/src/utils/phaseResultDisplay.tsx
@@ -12,6 +12,12 @@ import { Run } from "@/store/conversationStore";
 import { proseClasses } from "@/styles/proseClasses";
 import { updateRunUtil } from "@/utils/sendPrompts";
 import { useUserStore } from "@/store/userStore";
+import {
+  buildScoreBreakdown,
+  criterionHint,
+  getScoreColorByPct,
+  smoothScrollToElement,
+} from "@/utils/scoreDisplayUtils";
 
 interface AIResponseDisplayProps {
   run: Run | null;
@@ -25,202 +31,6 @@ interface RunScoreDisplayProps {
   explanationContent?: string | null;
 }
 
-type ScoreBreakdown = {
-  criteria: Array<{ name: string; score: number; max: number }>;
-  total: number;
-  totalMax: number;
-};
-
-function getScrollParent(el: HTMLElement): HTMLElement | Window {
-  let node: HTMLElement | null = el.parentElement;
-  while (node) {
-    const style = window.getComputedStyle(node);
-    const overflowY = style.overflowY;
-    const scrollable =
-      (overflowY === "auto" || overflowY === "scroll") &&
-      node.scrollHeight > node.clientHeight;
-    if (scrollable) return node;
-    node = node.parentElement;
-  }
-  return window;
-}
-
-function smoothScrollToElement(el: HTMLElement, durationMs = 850): () => void {
-  const TOP_OFFSET = 72;
-  const container = getScrollParent(el);
-  const startY =
-    container === window
-      ? window.scrollY
-      : (container as HTMLElement).scrollTop;
-  const targetY =
-    container === window
-      ? Math.max(0, window.scrollY + el.getBoundingClientRect().top - TOP_OFFSET)
-      : Math.max(
-          0,
-          (container as HTMLElement).scrollTop +
-            (el.getBoundingClientRect().top -
-              (container as HTMLElement).getBoundingClientRect().top) -
-            TOP_OFFSET
-        );
-  const distance = targetY - startY;
-  let rafId = 0;
-  const start = performance.now();
-
-  const easeInOutCubic = (t: number) =>
-    t < 0.5 ? 4 * t * t * t : 1 - Math.pow(-2 * t + 2, 3) / 2;
-
-  const step = (now: number) => {
-    const elapsed = now - start;
-    const p = Math.min(1, elapsed / durationMs);
-    const y = startY + distance * easeInOutCubic(p);
-    if (container === window) {
-      window.scrollTo(0, y);
-    } else {
-      (container as HTMLElement).scrollTop = y;
-    }
-    if (p < 1) {
-      rafId = window.requestAnimationFrame(step);
-    }
-  };
-
-  rafId = window.requestAnimationFrame(step);
-  return () => window.cancelAnimationFrame(rafId);
-}
-
-const SCORE_BAR_COLORS = {
-  red: "#DF3F46",
-  yellow: "#E3A135",
-  green: "#249953",
-};
-
-function getScoreColorByPct(pct: number): string {
-  if (pct <= 25) return SCORE_BAR_COLORS.red;
-  if (pct < 75) return SCORE_BAR_COLORS.yellow;
-  return SCORE_BAR_COLORS.green;
-}
-
-function extractJsonObject(text: string): Record<string, unknown> | null {
-  const trimmed = text.trim();
-  if (!trimmed) return null;
-  try {
-    const direct = JSON.parse(trimmed);
-    if (direct && typeof direct === "object" && !Array.isArray(direct)) {
-      return direct as Record<string, unknown>;
-    }
-  } catch {}
-
-  const fenceRe = /```(?:json)?\s*([\s\S]*?)\s*```/gi;
-  let m: RegExpExecArray | null;
-  while ((m = fenceRe.exec(trimmed)) !== null) {
-    try {
-      const parsed = JSON.parse(m[1].trim());
-      if (parsed && typeof parsed === "object" && !Array.isArray(parsed)) {
-        return parsed as Record<string, unknown>;
-      }
-    } catch {}
-  }
-
-  const firstBrace = trimmed.indexOf("{");
-  const lastBrace = trimmed.lastIndexOf("}");
-  if (firstBrace >= 0 && lastBrace > firstBrace) {
-    try {
-      const parsed = JSON.parse(trimmed.slice(firstBrace, lastBrace + 1));
-      if (parsed && typeof parsed === "object" && !Array.isArray(parsed)) {
-        return parsed as Record<string, unknown>;
-      }
-    } catch {}
-  }
-  return null;
-}
-
-function toNum(value: unknown): number | null {
-  if (typeof value === "number" && Number.isFinite(value)) return value;
-  if (typeof value === "string") {
-    const n = parseFloat(value.trim());
-    return Number.isNaN(n) ? null : n;
-  }
-  return null;
-}
-
-function parseRubricMaxMap(rubric?: string): Record<string, number> {
-  if (!rubric) return {};
-  const out: Record<string, number> = {};
-
-  try {
-    const parsed = JSON.parse(rubric) as unknown;
-    if (Array.isArray(parsed)) {
-      parsed.forEach((item) => {
-        if (!item || typeof item !== "object") return;
-        const criteria = (item as { criteria?: unknown }).criteria;
-        const lines = (item as { lines?: unknown }).lines;
-        if (typeof criteria !== "string" || !Array.isArray(lines)) return;
-
-        let max = 0;
-        lines.forEach((line) => {
-          if (!line || typeof line !== "object") return;
-          const s = toNum((line as { score?: unknown }).score);
-          if (s !== null && s > max) max = s;
-        });
-        if (max > 0) out[criteria] = max;
-      });
-      if (Object.keys(out).length > 0) return out;
-    }
-  } catch {}
-
-  const json = extractJsonObject(rubric);
-
-  if (json) {
-    Object.entries(json).forEach(([k, v]) => {
-      const n = toNum(v);
-      if (n !== null && n > 0) out[k] = n;
-    });
-  }
-  return out;
-}
-
-function buildScoreBreakdown(run: Run | null): ScoreBreakdown | null {
-  if (!run?.run_score) return null;
-  const scoreObj = extractJsonObject(run.run_score);
-  if (!scoreObj) return null;
-
-  const rubricMax = parseRubricMaxMap(run.scoreData?.rubric);
-  const criteria: Array<{ name: string; score: number; max: number }> = [];
-  let totalFromPayload: number | null = null;
-
-  Object.entries(scoreObj).forEach(([key, value]) => {
-    const n = toNum(value);
-    if (n === null) return;
-    if (key.toLowerCase() === "total") {
-      totalFromPayload = n;
-      return;
-    }
-    const max = rubricMax[key];
-    criteria.push({ name: key, score: n, max });
-  });
-
-  if (!criteria.length && totalFromPayload === null) return null;
-
-  const summed = criteria.reduce((acc, c) => acc + c.score, 0);
-  const total = totalFromPayload ?? summed;
-  const maxFromCriteria = criteria.reduce((acc, c) => acc + c.max, 0);
-  const totalMax =
-    maxFromCriteria > 0
-      ? maxFromCriteria
-      : Math.max(total, run.scoreData?.minimum_score || 0, 1);
-
-  return { criteria, total, totalMax };
-}
-
-function criterionHint(name: string, score: number, max: number): string {
-  const pct = max > 0 ? (score / max) * 100 : 0;
-  if (pct <= 25) {
-    return `Low score for ${name}. Focus on improving this area with clearer and more relevant points.`;
-  }
-  if (pct < 75) {
-    return `Decent ${name} score, but there is room to improve with stronger detail and structure.`;
-  }
-  return `Great ${name} performance. Keep using this level of quality in your response.`;
-}
 
 export const MarkdownResponseDisplay: React.FC<{
   content: string;
